backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
motor==3.3.1
python-dotenv==1.0.0
google-genai==1.65.0
python-multipart==0.0.6
pymupdf==1.23.6
